Split the module in two (read and write)
To support UEFI Secure Boot, the module should be split in half,
since in that mode GRUB must guarantee that writing to MSR is not
possible.

The new modules are called rdmsr and wrmsr, as the command that
each one implement.

Follow-up: http://lists.gnu.org/archive/html/grub-devel/2019-02/msg00072.html